+++ src/Files.App/MainWindow.xaml.cs
@@ -1,9 +1,7 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.App.CommandLine;
 using Files.App.UserControls.MultitaskingControl;
-using Files.App.Views;
 using Microsoft.UI;
 using Microsoft.UI.Windowing;
 using Microsoft.UI.Xaml.Controls;
