+++ src/Files.Backend/Enums/ParsedCommandType.cs
@@ -1,16 +1,22 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-namespace Files.App.CommandLine
+namespace Files.Backend.Enums
 {
-	internal enum ParsedCommandType
+	public enum ParsedCommandType
 	{
 		Unknown,
+
 		OpenDirectory,
+
 		OpenPath,
+
 		ExplorerShellCommand,
+
 		OutputPath,
+
 		SelectItem,
+
 		TagFiles
 	}
 }
